done algo and delete

+++ app/src/main/java/com/example/startracker/model/newImageModel.java
@@ -27,18 +27,20 @@ public class newImageModel {
         mDatabaseRefProcessed = FirebaseDatabase.getInstance().getReference("Users").child(id).child("processed");
     }
 
-    public void deleteImageModel( String refId, String refProcessedId, String storageId,String storageProcessedId, int flag){
-        mDatabaseRef.child(refId).removeValue()
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                    }
-                });
+    public void deleteImageModel( String refId, String refProcessedId, String storageId,String storageProcessedId, int flag, int flag1){
+        if(flag1==0){
+            mDatabaseRef.child(refId).removeValue()
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void aVoid) {
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                        }
+                    });
+        }
         mDatabaseRefProcessed.child(refProcessedId).removeValue()
                 .addOnSuccessListener(new OnSuccessListener<Void>() {
                     @Override
@@ -50,18 +52,20 @@ public class newImageModel {
                     public void onFailure(@NonNull Exception e) {
                     }
                 });
-        mStorageRef.child(storageId).delete()
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                   }
-                });
-        mStorageRef.child(storageProcessedId).delete()
+        if(flag1==0){
+            mStorageRef.child(storageId).delete()
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void aVoid) {
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                        }
+                    });
+        }
+        mStorageRefProcessed.child(storageProcessedId).delete()
                 .addOnSuccessListener(new OnSuccessListener<Void>() {
                     @Override
                     public void onSuccess(Void aVoid) {
